NODE-789-split-terms-and-types TYPES move from Terms to separate module.
