+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/compiler/CompilationError.scala
@@ -1,7 +1,7 @@
 package com.wavesplatform.lang.v1.compiler
 
 import cats.Show
-import com.wavesplatform.lang.v1.compiler.Terms.CASETYPEREF
+import com.wavesplatform.lang.v1.compiler.Types.CASETYPEREF
 import com.wavesplatform.lang.v1.evaluator.ctx.PredefFunction.FunctionTypeSignature
