+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/compiler/CompilerContext.scala
@@ -1,7 +1,7 @@
 package com.wavesplatform.lang.v1.compiler
 
 import com.wavesplatform.lang.v1.compiler.CompilerContext._
-import com.wavesplatform.lang.v1.compiler.Terms.TYPE
+import com.wavesplatform.lang.v1.compiler.Types.TYPE
 import com.wavesplatform.lang.v1.evaluator.ctx.PredefFunction.FunctionTypeSignature
 import com.wavesplatform.lang.v1.evaluator.ctx.{EvaluationContext, PredefBase}
 import shapeless._
