+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/compiler/CompilerV1.scala
@@ -9,6 +9,7 @@ import com.wavesplatform.lang.v1.FunctionHeader
 import com.wavesplatform.lang.v1.compiler.CompilationError._
 import com.wavesplatform.lang.v1.compiler.CompilerContext._
 import com.wavesplatform.lang.v1.compiler.Terms._
+import com.wavesplatform.lang.v1.compiler.Types._
 import com.wavesplatform.lang.v1.evaluator.ctx.PredefFunction.FunctionTypeSignature
 import com.wavesplatform.lang.v1.evaluator.ctx._
 import com.wavesplatform.lang.v1.evaluator.ctx.impl.PureContext
@@ -261,7 +262,7 @@ object CompilerV1 {
                        fieldName: String,
                        expr: EXPR): Either[CompilationError, (GETTER, TYPE)] = {
     types
-      .traverse[Option, Terms.TYPE](ctr => {
+      .traverse[Option, TYPE](ctr => {
         ctx.predefTypes
           .get(ctr.name)
           .flatMap(resolveFields(ctx, _).find(_._1 == fieldName).map(_._2))
