+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/compiler/TypeInferrer.scala
@@ -1,6 +1,6 @@
 package com.wavesplatform.lang.v1.compiler
 
-import com.wavesplatform.lang.v1.compiler.Terms._
+import com.wavesplatform.lang.v1.compiler.Types._
 import com.wavesplatform.lang._
 
 object TypeInferrer {
