+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/evaluator/ctx/Obj.scala
@@ -1,5 +1,5 @@
 package com.wavesplatform.lang.v1.evaluator.ctx
 
-import com.wavesplatform.lang.v1.compiler.Terms.CASETYPEREF
+import com.wavesplatform.lang.v1.compiler.Types.CASETYPEREF
 
 case class CaseObj(caseType: CASETYPEREF, fields: Map[String, Any])
