+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/testing/TypedScriptGen.scala
@@ -2,7 +2,7 @@ package com.wavesplatform.lang.v1.testing
 
 import com.wavesplatform.lang.v1.FunctionHeader
 import com.wavesplatform.lang.v1.compiler.Terms._
-import com.wavesplatform.lang.v1.compiler.Terms._
+import com.wavesplatform.lang.v1.compiler.Types._
 import org.scalacheck._
 import scodec.bits.ByteVector
 import com.wavesplatform.lang.v1.evaluator.FunctionIds._
